Color extension tests (#294)
* WIP testing

* Process feedback

* align tostring with MAUI parse method
* throw ArgumentOutOfRangeException
* removed WithAlpha
* cleanup test data

* Added tests WithCyan, Magenta, Yellow & BlackKey

* clean-up

* Migrate `ColorTestData.cs` to `ColorConversionExtensions_Tests.cs`

* Refactor ColorTestDefinition

+++ src/CommunityToolkit.Maui/Extensions/ColorAnimationExtensions.shared.cs
@@ -59,6 +59,6 @@ public static partial class ColorAnimationExtensions
 			new(v => element.BackgroundColor = element.BackgroundColor.WithBlue(v), element.BackgroundColor.Blue, targetBlue);
 
 		static Animation GetAlphaTransformAnimation(VisualElement element, float targetAlpha) =>
-			new(v => element.BackgroundColor = element.BackgroundColor.WithAlpha(v), element.BackgroundColor.Alpha, targetAlpha);
+			new(v => element.BackgroundColor = element.BackgroundColor.WithAlpha((float)v), element.BackgroundColor.Alpha, targetAlpha);
 	}
 }
